requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Task Queue
celery==5.3.6
redis==5.0.1

# OCR Libraries
paddleocr==2.7.3
paddlepaddle-gpu==2.6.0  # Use paddlepaddle for CPU-only
pillow==10.2.0
opencv-python==4.9.0.80
pdf2image==1.17.0
pytesseract==0.3.10

# Image Processing
numpy==1.26.3
scipy==1.12.0
scikit-image==0.22.0

# Document Processing
pypdf2==3.0.1
python-magic==0.4.27

# LLM Integration
transformers==4.37.2
torch==2.2.0
accelerate==0.26.1
sentencepiece==0.1.99

# API & Utilities
httpx==0.26.0
aiofiles==23.2.1
python-dateutil==2.8.2
pydantic-core==2.14.6

# Storage
boto3==1.34.34  # S3/MinIO
minio==7.2.3

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Security
cryptography==42.0.2

# Payment Processing
stripe==7.8.1

# Additional Utilities
python-magic-bin==0.4.14  # For Windows compatibility
